+++ src/Event/mt_evtLoadSave.c
@@ -44,7 +44,7 @@ void FUN_0038fa10(int param_1,u64 param_2);
 void FUN_0038ffb0(int param_1,int param_2);
 void FUN_003902c0(int param_1,int param_2);
 void FUN_003905f0(int param_1,int param_2);
-void FUN_00390920(int param_1,u64 param_2);
+void FUN_00390920(int param_1,int param_2);
 void FUN_00390c90(int param_1,u64 param_2);
 void FUN_00390ef0(int param_1,int param_2);
 void FUN_00391080(u64 param_1,u64 param_2);
@@ -3054,7 +3054,7 @@ void FUN_00390920(int param_1,int param_2)
 // FUN_00390C90 NONMATCHING
 
 
-void FUN_00390c90(int param_1,u64 param_2)
+void FUN_00390c90(int param_1,int param_2)
